[IMP] mail_compose_cc_bcc: split off account specific part
