--- mail_composer_cc_bcc/readme/ROADMAP.rst
@@ -1,1 +1,0 @@
-* Extract `account` customization (`account.invoice.send` wizard) to a specific module `mail_composer_cc_bcc_account`
