+++ mail_composer_cc_bcc_account/README.rst
@@ -1,13 +1,13 @@
-================
-Email CC and BCC
-================
+=====================================
+Email CC and BCC when sending invoice
+=====================================
 
 .. 
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:62de2f08eb7c85d6c08574eb92a22c35a42e4b66c34b120120c21fb49747edd8
+   !! source digest: sha256:8a4c8888e8b36687e3c4a17b49b957ccfbbe56e7b3227558ee5aef6ec447b164
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
@@ -17,10 +17,10 @@ Email CC and BCC
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsocial-lightgray.png?logo=github
-    :target: https://github.com/OCA/social/tree/16.0/mail_composer_cc_bcc
+    :target: https://github.com/OCA/social/tree/16.0/mail_composer_cc_bcc_account
     :alt: OCA/social
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/social-16-0/social-16-0-mail_composer_cc_bcc
+    :target: https://translation.odoo-community.org/projects/social-16-0/social-16-0-mail_composer_cc_bcc_account
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/builds?repo=OCA/social&target_branch=16.0
@@ -28,22 +28,8 @@ Email CC and BCC
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
 
-Odoo native does not support defining a Cc field in the Mail Composer
-by default; instead, it only has a unique Recipients fields, which is
-confusing for a lot of end users.
-
 This module allows to properly separate To:, Cc:, and Bcc: fields
-in the Mail Composer.
-
-Features
-~~~~~~~~
-
-* Add Cc and Bcc fields to mail composer form. Send only once to multiple email
-  addresses.
-* Add Cc and Bcc fields to company form to use them as default in mail composer
-  form.
-* Add Bcc field to mail template form. Use Cc and Bcc fields to lookup partners
-  by email then add them to corresponding fields in mail composer form.
+in when sending invoices.
 
 .. IMPORTANT::
    This is an alpha version, the data model and design can change at any time without warning.
@@ -58,28 +44,14 @@ Features
 Configuration
 =============
 
-In company form there are two fields to set default cc and bcc
-partners.
-
-  .. image:: https://raw.githubusercontent.com/OCA/social/16.0/mail_composer_cc_bcc/static/img/res_company_form_default_cc_bcc.png
-
-In template form there are two fields to set cc and bcc emails.
-
-  .. image:: https://raw.githubusercontent.com/OCA/social/16.0/mail_composer_cc_bcc/static/img/email_template_form_cc_bcc.png
+This module will be automatically installed if both account and
+mail_compose_cc_bcc are installed.
 
 Usage
 =====
 
-The partners cc and bcc from company form will be used to fill in mail composer
-form.
-
-  .. image:: https://raw.githubusercontent.com/OCA/social/16.0/mail_composer_cc_bcc/static/img/mail_compose_message_default_cc_bcc.png
-
-When select a template that has cc and/or bcc emails, the emails will be used
-to lookup partners then found partners will be added to corresponding mail
-composer's fields.
-
-  .. image:: https://raw.githubusercontent.com/OCA/social/16.0/mail_composer_cc_bcc/static/img/mail_compose_message_template_cc_bcc.png
+Use the cc and bcc fields when sending invoices, to notify other partners then
+the invoice contact of the customer.
 
 Bug Tracker
 ===========
@@ -87,7 +59,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/social/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_composer_cc_bcc%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_composer_cc_bcc_account%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -137,6 +109,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-hailangvn2023| 
 
-This module is part of the `OCA/social <https://github.com/OCA/social/tree/16.0/mail_composer_cc_bcc>`_ project on GitHub.
+This module is part of the `OCA/social <https://github.com/OCA/social/tree/16.0/mail_composer_cc_bcc_account>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
