+++ mail_composer_cc_bcc_account/__init__.py
@@ -1,0 +1,1 @@
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
