+++ mail_composer_cc_bcc_account/readme/CONFIGURE.rst
@@ -1,0 +1,2 @@
+This module will be automatically installed if both account and
+mail_compose_cc_bcc are installed.
