+++ mail_composer_cc_bcc_account/readme/CREDITS.rst
@@ -1,0 +1,1 @@
+The creation and migration from 15.0 to 16.0 of this module were financially supported by Camptocamp.
