+++ mail_composer_cc_bcc_account/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module allows to properly separate To:, Cc:, and Bcc: fields
+in when sending invoices.
