+++ mail_composer_cc_bcc_account/readme/USAGE.rst
@@ -1,0 +1,2 @@
+Use the cc and bcc fields when sending invoices, to notify other partners then
+the invoice contact of the customer.
