+++ mail_composer_cc_bcc_account/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_mail_cc_bcc
