+++ setup/mail_composer_cc_bcc_account/odoo/addons/mail_composer_cc_bcc_account
@@ -1,0 +1,1 @@
+../../../../mail_composer_cc_bcc_account
